+++ CHANGELOG.md
@@ -54,7 +54,9 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 	- `ImageInterface::toDataUri()`
 	- `ImageInterface::toStream()`
 	- `ImageInterface::getMimeType()`
-* Added `Scale` operation to the pixel library.
+* Added the following operations to the pixel library:
+	- `Polygon`
+	- `Scale`
 * Added operation `Pipeline` class to the pixel library.
 * It is now possible to implement custom image inspectors to retrieve information from the images.
 * Added `TopColors` inspector to the pixel library.